one user can now have access to multi app. An app can also be restricted to non-allowed users

+++ src/Dizda/Bundle/AppBundle/Request/PostDepositsRequest.php
@@ -32,8 +32,8 @@ class PostDepositsRequest extends AbstractRequest
         ));
 
         $resolver->setAllowedTypes(array(
-            'application_id' =>  ['integer'],
-            'type'           =>  ['integer'],
+            'application_id'  => ['numeric'],
+            'type'            => ['integer'],
             'amount_expected' => ['string'],
             'amount_expected_fiat' => ['array'],
             'reference'       => ['string', 'null']
